refactor: split metadata configuration modules
